+++ packages/vendure-plugin-accept-blue/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.0 (2024-03-07)
+
+- Always save payment methods and create recurring schedules and one time charges with a saved payment method
+- Allow payment for order with a saved payment method ID
+
 # 0.1.0 (2024-02-20)
 
 - Implement Charge Transaction with e2e tests